Struct async_compression::futures::write::ZstdDecoder

pub struct ZstdDecoder<W> { /* private fields */ }

Available on crate features futures-io and zstd only.

A zstd decoder, or decompressor

This structure implements an AsyncWrite interface and will take in compressed data and write it uncompressed to an underlying stream.

Implementations

impl<W: AsyncWrite> ZstdDecoder<W>

pub fn new(read: W) -> ZstdDecoder<W>

Creates a new decoder which will take in compressed data and write it uncompressed to the given stream.

pub fn with_params(inner: W, params: &[DParameter]) -> Self

Creates a new decoder, using the specified parameters, which will read compressed data from the given stream and emit a decompressed stream.

pub fn with_dict(inner: W, dictionary: &[u8]) -> Result<Self>

Creates a new decoder, using the specified compression level and pre-trained dictionary, which will read compressed data from the given stream and emit an uncompressed stream.

Dictionaries provide better compression ratios for small files, but are required to be present during decompression. The dictionary used must be the same as the one used for compression.

Errors

Returns error when dictionary is not valid.

pub fn get_ref(&self) -> &W

Acquires a reference to the underlying reader that this decoder is wrapping.

pub fn get_mut(&mut self) -> &mut W

Acquires a mutable reference to the underlying reader that this decoder is wrapping.

Note that care must be taken to avoid tampering with the state of the reader which may otherwise confuse this decoder.

pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut W>

Acquires a pinned mutable reference to the underlying reader that this decoder is wrapping.

Note that care must be taken to avoid tampering with the state of the reader which may otherwise confuse this decoder.

pub fn into_inner(self) -> W

Consumes this decoder returning the underlying reader.

Note that this may discard internal state of this decoder, so care should be taken to avoid losing resources when this is called.