Module async_compression::stream

Types which operate over Stream<Item = io::Result<Bytes>> streams, both encoders and decoders for various formats.

The Stream is treated as a single byte-stream to be compressed/decompressed, each item is a chunk of data from this byte-stream. There is not guaranteed to be a one-to-one relationship between chunks of data from the underlying stream and the resulting compressed/decompressed stream, the encoders and decoders will buffer the incoming data and choose their own boundaries at which to yield a new item.

Structs

BrotliDecoder
A brotli decoder, or decompressor.

BrotliEncoder
A brotli encoder, or compressor.

GzipDecoder
A gzip decoder, or decompressor.

GzipEncoder
A gzip encoder, or compressor.