Struct async_compression::tokio::write::buf_writer::BufWriter

pub struct BufWriter<W> {
    inner: W,
    buf: Box<[u8]>,
    written: usize,
    buffered: usize,
}

Fields
inner: Wbuf: Box<[u8]>written: usizebuffered: usize
Implementations

impl<W: AsyncWrite> BufWriter<W>

pub fn new(inner: W) -> Self

Creates a new BufWriter with a default buffer capacity. The default is currently 8 KB, but may change in the future.

pub fn with_capacity(cap: usize, inner: W) -> Self

Creates a new BufWriter with the specified buffer capacity.

fn partial_flush_buf( self: Pin<&mut Self>, cx: &mut Context<'_>, ) -> Poll<Result<()>>

fn flush_buf(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<()>>

impl<W> BufWriter<W>

pub fn get_ref(&self) -> &W

Gets a reference to the underlying writer.

pub fn get_mut(&mut self) -> &mut W

Gets a mutable reference to the underlying writer.

It is inadvisable to directly write to the underlying writer.

pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut W>

Gets a pinned mutable reference to the underlying writer.

It is inadvisable to directly write to the underlying writer.

pub fn into_inner(self) -> W

Consumes this BufWriter, returning the underlying writer.

Note that any leftover data in the internal buffer is lost.