Struct async_compression::tokio::write::BrotliDecoder

pub struct BrotliDecoder<W> { /* private fields */ }

A brotli decoder, or decompressor.

This structure implements an AsyncWrite interface and will take in compressed data and write it uncompressed to an underlying stream.

Implementations

impl<W: AsyncWrite> BrotliDecoder<W>

pub fn new(read: W) -> BrotliDecoder<W>

Creates a new decoder which will take in compressed data and write it uncompressedd to the given stream.

pub fn get_ref(&self) -> &W

Acquires a reference to the underlying reader that this decoder is wrapping.

pub fn get_mut(&mut self) -> &mut W

Acquires a mutable reference to the underlying reader that this decoder is wrapping.

Note that care must be taken to avoid tampering with the state of the reader which may otherwise confuse this decoder.

pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut W>

Acquires a pinned mutable reference to the underlying reader that this decoder is wrapping.

Note that care must be taken to avoid tampering with the state of the reader which may otherwise confuse this decoder.

pub fn into_inner(self) -> W

Consumes this decoder returning the underlying reader.

Note that this may discard internal state of this decoder, so care should be taken to avoid losing resources when this is called.