Crate async_compression

Adaptors between compression crates and Rust’s modern asynchronous IO types.

Feature Organization

This crate is divided up along two axes, which can each be individually selected via Cargo features.

All features are disabled by default, you should enable just the ones you need from the lists below.

If you want to pull in everything there are three group features defined:

Feature	Does

all	Activates all implementations and algorithms.

all-implementations	Activates all implementations, needs to be paired with a selection of algorithms

all-algorithms	Activates all algorithms, needs to be paired with a selection of implementations

IO implementation

The first division is which underlying asynchronous IO trait will be wrapped, these are available as separate features that have corresponding top-level modules:

Feature	Type

futures-io	futures::io::AsyncBufRead, futures::io::AsyncWrite

tokio (inactive)	tokio::io::AsyncBufRead, tokio::io::AsyncWrite

Compression algorithm

The second division is which compression schemes to support, there are currently a few available choices, these determine which types will be available inside the above modules:

Feature	Types

brotli (inactive)	BrotliEncoder, BrotliDecoder

bzip2 (inactive)	BzEncoder, BzDecoder

deflate (inactive)	DeflateEncoder, DeflateDecoder

gzip (inactive)	GzipEncoder, GzipDecoder

lzma (inactive)	LzmaEncoder, LzmaDecoder

xz (inactive)	XzEncoder, XzDecoder

zlib (inactive)	ZlibEncoder, ZlibDecoder

zstd	ZstdEncoder, ZstdDecoder

deflate64 (inactive)	(encoder not implemented), Deflate64Decoder

Modules

futuresfutures-io

Implementations for IO traits exported by futures-io.

zstdzstd

This module contains zstd-specific types for async-compression.

Enums

Level

Level of compression data should be compressed with.