Crate async_compression

Adaptors between compression crates and Rust’s modern asynchronous IO types.

Feature Organization

This crate is divided up along two axes, which can each be individually selected via Cargo features.

All features are disabled by default, you should enable just the ones you need from the lists below.

If you want to pull in everything there are three group features defined:

Feature	Does

all	Activates all implementations and algorithms.

all-implementations	Activates all implementations, needs to be paired with a selection of algorithms

all-algorithms	Activates all algorithms, needs to be paired with a selection of implementations

IO implementation

The first division is which underlying asynchronous IO trait will be wrapped, these are available as separate features that have corresponding top-level modules:

Feature	Type

futures-io (inactive)	futures::io::AsyncBufRead, futures::io::AsyncWrite

tokio	tokio::io::AsyncBufRead, tokio::io::AsyncWrite

Compression algorithm

The second division is which compression schemes to support, there are currently a few available choices, these determine which types will be available inside the above modules:

Feature	Types

brotli	BrotliEncoder, BrotliDecoder

bzip2 (inactive)	BzEncoder, BzDecoder

deflate (inactive)	DeflateEncoder, DeflateDecoder

gzip	GzipEncoder, GzipDecoder

lz4 (inactive)	Lz4Encoder, Lz4Decoder

lzma (inactive)	LzmaEncoder, LzmaDecoder

xz (inactive)	XzEncoder, XzDecoder

zlib (inactive)	ZlibEncoder, ZlibDecoder

zstd (inactive)	ZstdEncoder, ZstdDecoder

deflate64 (inactive)	(encoder not implemented), Deflate64Decoder

Multi-thread support

The xz compression algorithm supports multi-threaded compression and decompression. Enable the xz-parallel feature to enable multi-threading support.

Re-exports

pub use compression_codecs as codecs;

pub use compression_core as core;

Modules

brotli

This module contains Brotli-specific types for async-compression.

tokio

Implementations for IO traits exported by tokio v1.x.

Enums

Level

Level of compression data should be compressed with.