async_compression::generic::write::buf_write

Trait AsyncBufWrite

pub(crate) trait AsyncBufWrite {
    // Required method
    fn poll_partial_flush_buf(
        self: Pin<&mut Self>,
        cx: &mut Context<'_>,
    ) -> Poll<Result<Buffer<'_>>>;
}

Required Methods

fn poll_partial_flush_buf( self: Pin<&mut Self>, cx: &mut Context<'_>, ) -> Poll<Result<Buffer<'_>>>

Attempt to return an internal buffer to write to, flushing data out to the inner reader if it is full.

On success, returns Poll::Ready(Ok(buf)).

If the buffer is full and cannot be flushed, the method returns Poll::Pending and arranges for the current task context (cx) to receive a notification when the object becomes readable or is closed.